backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2
pinecone-client==3.0.2
httpx==0.26.0
python-multipart==0.0.6
